pyproject.toml
[project]
name = "signer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "sshkey-tools>=0.11.3",
    "click>=8.1.7",
    "asyncssh>=2.14.2",
    "fido2>=2.0.0",
    "pytimeparse>=1.1.8",
    "watchfiles>=1.1.0",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=8.18.1",
]
